:recycle: Split filters to dedicated component
