+++ issuepoker-frontend/src/components/IssueList.vue
@@ -9,112 +9,7 @@
     </v-col>
   </v-row>
   <v-row>
-    <v-col>
-      <v-row>
-        <v-col
-          cols="12"
-          sm="4"
-        >
-          <v-text-field
-            v-model="searchQuery"
-            :prepend-inner-icon="mdiMagnify"
-            clearable
-            density="compact"
-            hide-details
-            label="Suche (Titel, Beschreibung)"
-            variant="outlined"
-            @keyup.enter="search"
-            @click:clear="search"
-          />
-        </v-col>
-        <v-col
-          cols="12"
-          lg="2"
-          sm="4"
-        >
-          <v-autocomplete
-            v-model="filter.owners"
-            :items="filterOptions.owners"
-            :prepend-inner-icon="mdiHomeAccount"
-            chips
-            clearable
-            density="compact"
-            hide-details
-            label="Besitzer"
-            multiple
-            variant="outlined"
-            @update:model-value="fetchIssues"
-          />
-        </v-col>
-        <v-col
-          cols="12"
-          lg="2"
-          sm="4"
-        >
-          <v-autocomplete
-            v-model="filter.repositories"
-            :items="filterOptions.repositories"
-            :prepend-inner-icon="mdiSourceRepository"
-            chips
-            clearable
-            density="compact"
-            hide-details
-            label="Repository"
-            multiple
-            variant="outlined"
-            @update:model-value="fetchIssues"
-          />
-        </v-col>
-        <v-col
-          cols="12"
-          lg="2"
-          sm="4"
-        >
-          <v-select
-            v-model="filter.voted"
-            :items="selectOptions"
-            :prepend-inner-icon="mdiVote"
-            density="compact"
-            hide-details
-            label="Abgestimmt"
-            variant="outlined"
-            @update:model-value="fetchIssues"
-          />
-        </v-col>
-        <v-col
-          cols="12"
-          lg="2"
-          sm="4"
-        >
-          <v-select
-            v-model="filter.resulted"
-            :items="selectOptions"
-            :prepend-inner-icon="mdiTrophy"
-            density="compact"
-            hide-details
-            label="Ergebnis"
-            variant="outlined"
-            @update:model-value="fetchIssues"
-          />
-        </v-col>
-      </v-row>
-    </v-col>
-    <v-col cols="auto">
-      <v-tooltip
-        location="top"
-        text="Filter zurücksetzen"
-      >
-        <template v-slot:activator="{ props }">
-          <v-btn
-            :icon="mdiRestore"
-            density="comfortable"
-            rounded="rounded"
-            v-bind="props"
-            @click="resetFilters"
-          />
-        </template>
-      </v-tooltip>
-    </v-col>
+    <issue-list-filters @fetchIssues="fetchIssues" />
   </v-row>
   <v-row>
     <v-data-table-server
@@ -126,7 +21,7 @@
       :items-per-page-options="itemsPerPageOptions"
       :loading="loading"
       :page="page"
-      :search="searchQuery"
+      :search="getSearchQuery"
       :sort-by="sortBy"
       multi-sort
       @update:options="updateOptions"
@@ -136,38 +31,27 @@
 </template>
 
 <script lang="ts" setup>
-import type { FilterOptions } from "@/types/FilterOptions.ts";
 import type IssueDetails from "@/types/IssueDetails.ts";
 import type IssueSummary from "@/types/IssueSummary.ts";
 import type Page from "@/types/Page.ts";
 import type { SortItem } from "vuetify/lib/components/VDataTable/composables/sort.js";
 
-import {
-  mdiHomeAccount,
-  mdiMagnify,
-  mdiRestore,
-  mdiSourceRepository,
-  mdiTrophy,
-  mdiVote,
-} from "@mdi/js";
 import { storeToRefs } from "pinia";
-import { computed, onMounted, ref } from "vue";
+import { ref } from "vue";
 
-import { getFilterOptions } from "@/api/fetch-filterOptions.ts";
 import { getIssueList } from "@/api/fetch-issueList.ts";
+import IssueListFilters from "@/components/IssueListFilters.vue";
 import {
   ROLE_ADMIN,
-  ROUTES_HOME,
   ROUTES_ISSUE_CREATE,
   ROUTES_ISSUE_DETAIL,
 } from "@/constants.ts";
 import router from "@/plugins/router.ts";
-import { defaultFilter, useFilterStore } from "@/stores/filter.ts";
+import { useFilterStore } from "@/stores/filter.ts";
 import { useSearchQueryStore } from "@/stores/searchQuery.ts";
 import { useSnackbarStore } from "@/stores/snackbar.ts";
 import { useUserStore } from "@/stores/user.ts";
 
-const snackbarStore = useSnackbarStore();
 const headers = [
   { key: "owner", title: "Besitzer" },
   { key: "repository", title: "Repository" },
@@ -184,15 +68,14 @@ const itemsPerPageOptions = [
   { value: 100, title: "100" },
   { value: -1, title: "$vuetify.dataFooter.itemsPerPageAll" },
 ];
-const selectOptions = [
-  { value: null, title: "Egal" },
-  { value: true, title: "Ja" },
-  { value: false, title: "Nein" },
-];
 
+const snackbarStore = useSnackbarStore();
 const { getUser } = storeToRefs(useUserStore());
-const issues = ref<IssueSummary[]>([]);
+const { getSearchQuery } = storeToRefs(useSearchQueryStore());
+const { getFilter } = storeToRefs(useFilterStore());
+
 const loading = ref(true);
+const issues = ref<IssueSummary[]>([]);
 const totalIssues = ref(0);
 const page = ref<number>(1);
 const itemsPerPage = ref<number>(10);
@@ -201,30 +84,6 @@ const sortBy = ref<SortItem[]>([
   { key: "repository", order: "asc" },
   { key: "number", order: "asc" },
 ]);
-const searchQueryStore = useSearchQueryStore();
-const searchQuery = computed({
-  get: () => searchQueryStore.getSearchQuery,
-  set: (value) => {
-    searchQueryStore.setSearchQuery(value);
-  },
-});
-const filterOptions = ref<FilterOptions>(defaultFilter());
-const filterStore = useFilterStore();
-const filter = computed({
-  get: () => filterStore.getFilter,
-  set: (value) => {
-    filterStore.setFilter(value);
-    fetchIssues();
-  },
-});
-
-onMounted(() => {
-  getFilterOptions()
-    .then((content: FilterOptions) => (filterOptions.value = content))
-    .catch((error) => {
-      snackbarStore.showMessage(error);
-    });
-});
 
 function fetchIssues() {
   loading.value = true;
@@ -232,8 +91,8 @@ function fetchIssues() {
     page.value - 1,
     itemsPerPage.value,
     sortBy.value,
-    searchQuery.value ?? "",
-    filter.value
+    getSearchQuery.value || "",
+    getFilter.value
   )
     .then((content: Page<IssueSummary>) => {
       issues.value = content.content;
@@ -272,20 +131,5 @@ function goToIssue(
       number: props.item.number,
     },
   });
-}
-
-function search() {
-  searchQueryStore.setSearchQuery(searchQuery.value);
-  if (searchQuery.value && searchQuery.value != "") {
-    router.push({ name: ROUTES_HOME, query: { search: searchQuery.value } });
-  } else {
-    router.push({ name: ROUTES_HOME });
-  }
-}
-
-function resetFilters() {
-  filterStore.setFilter(defaultFilter());
-  searchQueryStore.setSearchQuery("");
-  fetchIssues();
 }
 </script>
